Add ownership checks to portfolios

+++ src/portfolios/domain/portfolios.service.ts
@@ -1,4 +1,9 @@
-import { Injectable, Logger, NotFoundException } from '@nestjs/common';
+import {
+  Injectable,
+  Logger,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { Cron } from '@nestjs/schedule';
 import { isValid } from 'date-fns';
 import { isNumber } from 'lodash';
@@ -49,16 +54,15 @@ export class PortfoliosService {
       : this.repository.findByOwnerId(user.id);
   }
 
-  async findOne(uuid: string): Promise<PortfolioDetailDto> {
-    //TODO: check ownership
+  async findOne(user: User, uuid: string): Promise<PortfolioDetailDto> {
     const portfolio = await this.repository.findOne(uuid);
-
     if (!portfolio) {
       throw new NotFoundException('Portfolio not found');
     }
+    this.checkOwner(user, portfolio);
 
     const positions =
-      await this.positionService.getPositionDetailsByPortfolioUuid(uuid);
+      await this.positionService.getPositionDetailsByPortfolioUuid(portfolio);
     const state = await this.portfolioStatesService.getLastByPortfolioUuid(
       uuid,
     );
@@ -75,28 +79,26 @@ export class PortfoliosService {
     };
   }
 
-  async deleteOne(uuid: string) {
-    //TODO: check ownership
+  async deleteOne(user: User, uuid: string) {
     const portfolio = await this.repository.findOne(uuid);
-
     if (!portfolio) {
       throw new NotFoundException('Portfolio not found');
     }
+    this.checkOwner(user, portfolio);
 
-    await this.positionService.deleteByPortfolioUuid(uuid);
+    await this.positionService.deleteByPortfolioUuid(user, uuid);
     await this.portfolioStatesService.deleteByPortfolioUuid(uuid);
     await this.repository.deleteOne(uuid);
 
     return portfolio;
   }
 
-  async getAverageBalances(uuid: string, range: string) {
-    //TODO: check ownership
+  async getAverageBalances(user: User, uuid: string, range: string) {
     const portfolio = await this.repository.findOne(uuid);
-
     if (!portfolio) {
       throw new NotFoundException('Portfolio not found');
     }
+    this.checkOwner(user, portfolio);
 
     return this.portfolioStatesService.getAverageBalancesForRange(
       uuid,
@@ -105,14 +107,15 @@ export class PortfoliosService {
   }
 
   async updateCash(
+    user: User,
     portfolioUuid: string,
     updatePortfolioCashDto: UpdatePortfolioCashDto,
   ): Promise<Portfolio> {
-    //TODO: check ownership
     const portfolio = await this.repository.findOne(portfolioUuid);
     if (!portfolio) {
       throw new NotFoundException(`Portfolio not found`);
     }
+    this.checkOwner(user, portfolio);
 
     // TODO: implement validation
     const { cash } = updatePortfolioCashDto;
@@ -127,14 +130,15 @@ export class PortfoliosService {
   }
 
   async addContribution(
+    user: User,
     uuid: string,
     addPortfolioContributionDto: AddPortfolioContributionDto,
   ): Promise<Portfolio> {
-    //TODO: check ownership
     const portfolio = await this.repository.findOne(uuid);
     if (!portfolio) {
       throw new NotFoundException(`Portfolio not found`);
     }
+    this.checkOwner(user, portfolio);
 
     // TODO: implement validation
     const { timestamp: inputTS, amountEUR } = addPortfolioContributionDto;
@@ -154,19 +158,27 @@ export class PortfoliosService {
   }
 
   async deleteContribution(
+    user: User,
     portfolioUuid: string,
     contributionUuid: string,
   ): Promise<Portfolio> {
-    //TODO: check ownership
     const portfolio = await this.repository.findOne(portfolioUuid);
+
     if (!portfolio) {
       throw new NotFoundException(`Portfolio not found`);
     }
+    this.checkOwner(user, portfolio);
 
     await this.repository.deleteContribution(portfolioUuid, contributionUuid);
     const updated = await this.repository.findOne(portfolioUuid);
     await this.positionService.updatePortfolioState(updated);
     return updated;
+  }
+
+  private checkOwner(user: User, portfolio: Portfolio): void {
+    if (portfolio.ownerId !== user.id) {
+      throw new UnauthorizedException('Access denied');
+    }
   }
 
   @Cron('0 0 8 * * *', { timeZone: 'Europe/Madrid' })
